* 'feature-B.3.1-jb' of https://github.com/K-Cassie/IS4103KMS: (26 commits)
  add view all project component html
  add create new project component html
  fix stack overflow
  edit project user relationship
  Merge branch 'master' into feature-B.1
  make edit sdg clickable and include follow request sent button in profile page
  fixed follow user for private account, get add remove affiliated user backend
  add lat lng to mras
  fixed add remove SDG
  Add and remove sdg
  fix select2
  add view followers and following
  reset password and inactivity logout
  Report user works
  report profile
  fixed search users
  Delete User (backend) retrieve User (backend)
  report profile backend
  add share profile and create toast for duplicate follow request
  changed get all users method, updated UI of search user
  ...

# Conflicts:
#	KMS/KMS-war/src/java/ws/restful/resources/ApplicationConfig.java
#	KMSAngular/package.json

+++ KMS/KMS-war/src/java/ejb/session/stateless/UserSessionBean.java
@@ -1,21 +1,35 @@
 package ejb.session.stateless;
 
+import Exception.AffiliatedUserExistException;
 import Exception.DuplicateEmailException;
+import Exception.DuplicateFollowRequestException;
 import Exception.DuplicateTagInProfileException;
 import Exception.InvalidLoginCredentialException;
+import Exception.InvalidUUIDException;
 import Exception.NoResultException;
 import Exception.UserNotFoundException;
 import entity.FollowRequestEntity;
-import entity.MaterialResourceAvailableEntity;
 import entity.TagEntity;
 import entity.UserEntity;
+import java.util.ArrayList;
+import java.security.SecureRandom;
 import java.util.List;
+import java.util.Properties;
+import java.util.UUID;
 import javax.ejb.Stateless;
+import javax.mail.Message;
+import javax.mail.MessagingException;
+import javax.mail.PasswordAuthentication;
+import javax.mail.Session;
+import javax.mail.Transport;
+import javax.mail.internet.InternetAddress;
+import javax.mail.internet.MimeMessage;
 import javax.persistence.EntityManager;
 import javax.persistence.NonUniqueResultException;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
 import util.enumeration.AccountPrivacySettingEnum;
+import util.enumeration.UserTypeEnum;
 import util.security.CryptographicHelper;
 
 /**
@@ -36,9 +50,75 @@ public class UserSessionBean implements UserSessionBeanLocal {
         if (!q.getResultList().isEmpty()) {
             throw new DuplicateEmailException("Email already exist!");
         }
+        String uuid = UUID.randomUUID().toString().replace("-", "");
+        user.setVerificationCode(uuid);
+        user.setIsVerified(Boolean.FALSE);
         em.persist(user);
         em.flush();
         return user;
+    }
+    
+    @Override
+    public void resetPassword(String email)throws UserNotFoundException{
+        String password = this.generateRandomPassword(10);
+        UserEntity user = this.retrieveUserByEmail(email);
+        user.setPassword(password);
+        em.merge(user);
+        em.flush();
+        this.sendResetPasswordEmail(user.getEmail(), password);
+    }
+    
+    private String generateRandomPassword(int len){
+        
+        final String chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+        SecureRandom random = new SecureRandom();
+        StringBuilder sb = new StringBuilder();
+        
+        for (int i = 0; i < len; i++){
+            int randomIndex = random.nextInt(chars.length());
+            sb.append(chars.charAt(randomIndex));
+        }
+        return sb.toString();
+    }
+    
+    private void sendResetPasswordEmail(String destinationEmail, String newPassword){
+        
+        final String username = "4103kms";
+        final String password = "4103kmsemail";
+        final String host = "localhost";
+        
+        Properties props = new Properties();
+        props.put("mail.smtp.auth", "true");
+        props.put("mail.smtp.starttls.enable", "true");
+        props.put("mail.smtp.host", "smtp.gmail.com");
+        props.put("mail.smtp.port", "587");
+        props.put("mail.smtp.ssl.trust", "smtp.gmail.com");
+        props.put("mail.smtp.user", username);
+        
+        Session session = Session.getInstance(props, new javax.mail.Authenticator() {
+            protected PasswordAuthentication getPasswordAuthentication(){
+                return new PasswordAuthentication(username, password);
+            }
+        });
+        
+        try{
+            Message message = new MimeMessage(session);
+            message.setFrom(new InternetAddress("4103kms@gmail.com"));
+            message.setRecipients(Message.RecipientType.TO, InternetAddress.parse(destinationEmail));
+            message.setSubject("Password Reset for KMS");
+            message.setText("Dear user, "+ 
+                    "\n" +
+                    "\n" + "Your password has been reset. New password: " + newPassword +
+                    "\n" + "Please login and change your password");
+            
+            Transport.send(message);
+            
+            System.out.println("reset password message sent");
+        }
+        catch(MessagingException ex){
+            throw new RuntimeException(ex);
+        }
+
     }
 
     @Override
@@ -94,7 +174,9 @@ public class UserSessionBean implements UserSessionBeanLocal {
             user.getGroupsJoined().size();
             user.getGroupsOwned().size();
             user.getPosts().size();
-            //user.getProjects().size();
+            user.getProjectAdmins().size();
+            user.getProjectsJoined().size();
+            user.getProjectsOwned().size();
             user.getReviewsGiven().size();
             user.getSdgs().size();
             user.getSkills().size();
@@ -120,20 +202,25 @@ public class UserSessionBean implements UserSessionBeanLocal {
             String passwordHash = CryptographicHelper.getInstance().byteArrayToHexString(CryptographicHelper.getInstance().doMD5Hashing(password + user.getSalt()));
             if (user.getPassword().equals(passwordHash)) {
                 System.out.println("user login if()");
-                System.out.println(user.getPassword());
-                System.out.println(passwordHash);
                 return user;
             } else {
                 System.out.println("user login else()");
-                System.out.println(user.getPassword());
-                System.out.println(passwordHash);
                 throw new InvalidLoginCredentialException("Email does not exist of invalid password");
             }
         } catch (UserNotFoundException ex) {
-            System.out.println("here");
             throw new InvalidLoginCredentialException("Email does not exist or invalid password!");
         }
 
+    }
+
+    @Override
+    public List<TagEntity> getSDGsForProfile(long userId) throws UserNotFoundException {
+        UserEntity user = em.find(UserEntity.class, userId);
+
+        if (user == null) {
+            throw new UserNotFoundException("User not found.");
+        }
+        return user.getSdgs();
     }
 
     @Override
@@ -167,6 +254,60 @@ public class UserSessionBean implements UserSessionBeanLocal {
         sdgs.remove(tag);
         user.setSdgs(sdgs);
     }
+    
+    @Override
+    public void sendVerificationEmail(String destinationEmail, String verificationCode){
+        
+        final String username = "4103kms";
+        final String password = "4103kmsemail";
+        final String host = "localhost";
+        
+        Properties props = new Properties();
+        props.put("mail.smtp.auth", "true");
+        props.put("mail.smtp.starttls.enable", "true");
+        props.put("mail.smtp.host", "smtp.gmail.com");
+        props.put("mail.smtp.port", "587");
+        props.put("mail.smtp.ssl.trust", "smtp.gmail.com");
+        props.put("mail.smtp.user", username);
+        
+        Session session = Session.getInstance(props, new javax.mail.Authenticator() {
+            protected PasswordAuthentication getPasswordAuthentication(){
+                return new PasswordAuthentication(username, password);
+            }
+        });
+        
+        try{
+            Message message = new MimeMessage(session);
+            message.setFrom(new InternetAddress("4103kms@gmail.com"));
+            message.setRecipients(Message.RecipientType.TO, InternetAddress.parse(destinationEmail));
+            message.setSubject("Thank you for signing up with kms");
+            message.setText("Please verify your email address at http://localhost:4200//accountVerification/" + destinationEmail + "/"+ verificationCode);
+            
+            Transport.send(message);
+            
+            System.out.println("message sent");
+        }
+        catch(MessagingException ex){
+            throw new RuntimeException(ex);
+        }
+
+    }
+    
+    public Boolean verifyEmail(String email, String uuid) throws UserNotFoundException, InvalidUUIDException{
+        
+        UserEntity user = this.retrieveUserByEmail(email);
+        if(user.getIsVerified() == Boolean.TRUE){
+            return user.getIsVerified();
+        }
+        
+        if(user.getVerificationCode().equals(uuid)){
+            user.setIsVerified(Boolean.TRUE);
+            em.flush();
+        }else{
+            throw new InvalidUUIDException("Invalid UUID");
+        }
+        return user.getIsVerified();
+    }
 
     @Override
     public void deleteUser(long userId, UserEntity user) throws NoResultException {
@@ -176,9 +317,7 @@ public class UserSessionBean implements UserSessionBeanLocal {
     }
 
     @Override
-    public void followUser(Long toUserId, Long fromUserId) throws UserNotFoundException {
-        System.out.println(toUserId);
-        System.out.println(fromUserId);
+    public FollowRequestEntity followUser(Long toUserId, Long fromUserId) throws UserNotFoundException, DuplicateFollowRequestException {
         UserEntity toUser = em.find(UserEntity.class, toUserId);
         UserEntity fromUser = em.find(UserEntity.class, fromUserId);
         if (toUser == null || fromUser == null) {
@@ -187,31 +326,74 @@ public class UserSessionBean implements UserSessionBeanLocal {
         if (toUser.getAccountPrivacySetting().equals(AccountPrivacySettingEnum.PUBLIC)) {
             fromUser.getFollowing().add(toUser);
             toUser.getFollowers().add(fromUser);
+            return null;
         } else {
-            FollowRequestEntity followRequestEntity = new FollowRequestEntity();
-            em.persist(followRequestEntity);
-            followRequestEntity.setFrom(fromUser);
-            followRequestEntity.setTo(toUser);
+            Query q = em.createQuery("SELECT f FROM FollowRequestEntity AS f WHERE f.from.userId = :from AND f.to.userId = :to");
+            q.setParameter("from", fromUserId);
+            q.setParameter("to", toUserId);
+            try {
+                FollowRequestEntity f = (FollowRequestEntity) q.getSingleResult();
+                throw new DuplicateFollowRequestException("Follow request already sent!");
+            } catch (javax.persistence.NoResultException e) {
+                FollowRequestEntity followRequestEntity = new FollowRequestEntity();
+                followRequestEntity.setFrom(fromUser);
+                followRequestEntity.setTo(toUser);
+
+                fromUser.getFollowRequestMade().add(followRequestEntity);
+                toUser.getFollowRequestReceived().add(followRequestEntity);
+                em.persist(followRequestEntity);
+                em.flush();
+                return followRequestEntity;
+            }
         }
     }
 
     @Override
     public void acceptFollowRequest(Long toUserId, Long fromUserId) throws NoResultException, UserNotFoundException {
-        Query q = em.createQuery("SELECT f FROM FollowRequestEntity AS f WHERE f.from=:from AND f.to=:to");
+        Query q = em.createQuery("SELECT f FROM FollowRequestEntity AS f WHERE f.from.userId = :from AND f.to.userId = :to");
+        q.setParameter("from", fromUserId);
+        q.setParameter("to", toUserId);
+        FollowRequestEntity f = (FollowRequestEntity) q.getSingleResult();
+        if (f == null) {
+            throw new NoResultException("Follow request not found");
+        }
+        UserEntity toUser = em.find(UserEntity.class, toUserId);
+        UserEntity fromUser = em.find(UserEntity.class, fromUserId);
+        if (toUser == null || fromUser == null) {
+            throw new UserNotFoundException("User not found");
+        }
+        fromUser.getFollowRequestMade().remove(f);
+        toUser.getFollowRequestReceived().remove(f);
+        f.setTo(null);
+        f.setFrom(null);
+        em.remove(f);
+        if (!fromUser.getFollowing().contains(toUser)) {
+            fromUser.getFollowing().add(toUser);
+        }
+        if (!toUser.getFollowers().contains(fromUser)) {
+            toUser.getFollowers().add(fromUser);
+        }
+    }
+
+    @Override
+    public void rejectFollowRequest(Long toUserId, Long fromUserId) throws NoResultException, UserNotFoundException {
+        Query q = em.createQuery("SELECT f FROM FollowRequestEntity AS f WHERE f.from.userId = :from AND f.to.userId = :to");
         q.setParameter("from", fromUserId);
         q.setParameter("to", toUserId);
         FollowRequestEntity f = (FollowRequestEntity) q.getSingleResult();
         if (f == null) {
             throw new NoResultException("follow request not found");
         }
-        em.remove(f);
         UserEntity toUser = em.find(UserEntity.class, toUserId);
-        UserEntity fromUser = em.find(UserEntity.class, toUserId);
+        UserEntity fromUser = em.find(UserEntity.class, fromUserId);
         if (toUser == null || fromUser == null) {
             throw new UserNotFoundException("User not found");
         }
-        fromUser.getFollowing().add(toUser);
-        toUser.getFollowers().add(fromUser);
+        fromUser.getFollowRequestMade().remove(f);
+        toUser.getFollowRequestReceived().remove(f);
+        f.setTo(null);
+        f.setFrom(null);
+        em.remove(f);
     }
 
     @Override
@@ -250,14 +432,80 @@ public class UserSessionBean implements UserSessionBeanLocal {
         return skillTags;
     }
 
+
     @Override
-    public List<UserEntity> getAllUsers() throws NoResultException{
+    public List<UserEntity> getAllUsers() throws NoResultException {
         Query q = em.createQuery("SELECT u FROM UserEntity u");
         List<UserEntity> users = q.getResultList();
+        for (UserEntity userEntity : users) {
+            userEntity.getFollowers().size();
+            userEntity.getFollowing().size();
+        }
         return users;
     }
+
+    @Override
+    public List<UserEntity> getAffiliatedUsers(Long userId) throws UserNotFoundException {
+        UserEntity user = em.find(UserEntity.class, userId);
+        if (user == null) {
+            throw new UserNotFoundException("User not found");
+        }
+        List<UserEntity> users = new ArrayList<>();
+        if (user.getUserType() == UserTypeEnum.INDIVIDUAL) {
+            users = user.getAffiliatedInstitutes();
+        } else if (user.getUserType() == UserTypeEnum.INSTITUTE) {
+            users = user.getAffiliatedIndividuals();
+        }
+        for (UserEntity userEntity : users) {
+            userEntity.getFollowers().size();
+            userEntity.getFollowing().size();
+        }
+        return users;
+    }
+
+    @Override
+    public void addAffiliatedUser(Long userId, Long affiliatedToAddUserId) throws AffiliatedUserExistException, UserNotFoundException {
+        UserEntity user = em.find(UserEntity.class, userId);
+        UserEntity affiliatedUserToAdd = em.find(UserEntity.class, affiliatedToAddUserId);
+        if (user == null || affiliatedUserToAdd == null) {
+            throw new UserNotFoundException("User not found");
+        }
+        List<UserEntity> affiliatedUsers = new ArrayList<>();
+        if (user.getUserType() == UserTypeEnum.INDIVIDUAL) {
+            affiliatedUsers = user.getAffiliatedInstitutes();
+        } else if (user.getUserType() == UserTypeEnum.INSTITUTE) {
+            affiliatedUsers = user.getAffiliatedIndividuals();
+        }
+
+        if (!affiliatedUsers.contains(affiliatedUserToAdd)) {
+            affiliatedUsers.add(affiliatedUserToAdd);
+        } else{
+            throw new AffiliatedUserExistException("User is already affiliated.");
+        }
+    }
     
-    public UserEntity updateUser(UserEntity updatedUser) throws UserNotFoundException, DuplicateEmailException {
+    @Override
+    public void removeAffiliatedUser(Long userId, Long affiliatedToRemoveUserId) throws NoResultException, UserNotFoundException {
+        UserEntity user = em.find(UserEntity.class, userId);
+        UserEntity affiliatedUserToRemove = em.find(UserEntity.class, affiliatedToRemoveUserId);
+        if (user == null || affiliatedUserToRemove == null) {
+            throw new UserNotFoundException("User not found");
+        }
+        List<UserEntity> affiliatedUsers = new ArrayList<>();
+        if (user.getUserType() == UserTypeEnum.INDIVIDUAL) {
+            affiliatedUsers = user.getAffiliatedInstitutes();
+        } else if (user.getUserType() == UserTypeEnum.INSTITUTE) {
+            affiliatedUsers = user.getAffiliatedIndividuals();
+        }
+
+        if (affiliatedUsers.contains(affiliatedUserToRemove)) {
+            affiliatedUsers.remove(affiliatedUserToRemove);
+        } else{
+            throw new NoResultException("User is not affiliated.");
+        }
+    }
+
+    public UserEntity updateUser(UserEntity updatedUser) throws UserNotFoundException, DuplicateEmailException, NoResultException {
         UserEntity user = em.find(UserEntity.class, updatedUser.getUserId());
         if (user == null) {
             throw new UserNotFoundException("User not found");
@@ -276,6 +524,14 @@ public class UserSessionBean implements UserSessionBeanLocal {
         user.setProfilePicture(updatedUser.getProfilePicture());
         user.setAccountPrivacySetting(updatedUser.getAccountPrivacySetting());
 
+        for (int i = 0; i < updatedUser.getSdgs().size(); i++) {
+            TagEntity tag = em.find(TagEntity.class, updatedUser.getSdgs().get(i).getTagId());
+            if (tag == null) {
+                throw new NoResultException("SDG tag not found.");
+            }
+        }
+        user.setSdgs(updatedUser.getSdgs());
+
         return user;
     }
 
@@ -285,24 +541,37 @@ public class UserSessionBean implements UserSessionBeanLocal {
         if (user == null) {
             throw new UserNotFoundException("User not found");
         }
+        user.getFollowers().size();
         return user.getFollowers();
     }
-    
+
     @Override
     public List<UserEntity> getFollowing(long userId) throws UserNotFoundException {
         UserEntity user = em.find(UserEntity.class, userId);
         if (user == null) {
             throw new UserNotFoundException("User not found");
         }
+        user.getFollowing().size();
         return user.getFollowing();
     }
-    
+
     @Override
-    public List<MaterialResourceAvailableEntity> getMaterialRequestAvailable(long userId) throws UserNotFoundException {
+    public List<FollowRequestEntity> getFollowRequestReceived(Long userId) throws UserNotFoundException {
         UserEntity user = em.find(UserEntity.class, userId);
         if (user == null) {
             throw new UserNotFoundException("User not found");
         }
-        return user.getMras();
+        user.getFollowRequestReceived().size();
+        return user.getFollowRequestReceived();
+    }
+
+    @Override
+    public List<FollowRequestEntity> getFollowRequestMade(Long userId) throws UserNotFoundException {
+        UserEntity user = em.find(UserEntity.class, userId);
+        if (user == null) {
+            throw new UserNotFoundException("User not found");
+        }
+        user.getFollowRequestMade().size();
+        return user.getFollowRequestMade();
     }
 }
